add: defaultとadminとで共通コンポーネントを動的に出仕分けるよう変更
