+++ front/src/components/layouts/Admin.vue
@@ -1,0 +1,3 @@
+<template>
+  admindesu
+</template>
